move _getfbzeros to fb.py
